resume-rag-voice-agent/backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
supabase
python-dotenv
openai
elevenlabs
pypdf
mcp
websockets
python-multipart
